# Optimize number parsing