--- a/src/System.Private.CoreLib/shared/System/Globalization/NumberFormatInfo.cs
+++ b/src/System.Private.CoreLib/shared/System/Globalization/NumberFormatInfo.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-using System.Text;
 
 namespace System.Globalization
 {
@@ -39,16 +37,11 @@ namespace System.Globalization
     // CurrencySymbol            "$"      String used as local monetary symbol.
     //
 
-    sealed public class NumberFormatInfo : IFormatProvider, ICloneable
+    public sealed class NumberFormatInfo : IFormatProvider, ICloneable
     {
         // invariantInfo is constant irrespective of your current culture.
         private static volatile NumberFormatInfo s_invariantInfo;
 
-        // READTHIS READTHIS READTHIS
-        // This class has an exact mapping onto a native structure defined in COMNumber.cpp
-        // DO NOT UPDATE THIS WITHOUT UPDATING THAT STRUCTURE. IF YOU ADD BOOL, ADD THEM AT THE END.
-        // ALSO MAKE SURE TO UPDATE mscorlib.h in the VM directory to check field offsets.
-        // READTHIS READTHIS READTHIS
         internal int[] numberGroupSizes = new int[] { 3 };
         internal int[] currencyGroupSizes = new int[] { 3 };
         internal int[] percentGroupSizes = new int[] { 3 };
@@ -83,12 +76,9 @@ namespace System.Globalization
 
         internal bool isReadOnly = false;
 
-        // Is this NumberFormatInfo for invariant culture?
-        internal bool m_isInvariant = false;
+        private bool _hasInvariantNumberSigns = true;
 
-        internal bool HasInvariantNumberSigns { get; private set; } = true;
-
-        public NumberFormatInfo() : this(null)
+        public NumberFormatInfo()
         {
         }
 
@@ -173,9 +163,11 @@ namespace System.Globalization
             }
         }
 
+        internal bool HasInvariantNumberSigns => _hasInvariantNumberSigns;
+
         private void UpdateHasInvariantNumberSigns()
         {
-            HasInvariantNumberSigns = positiveSign == "+" && negativeSign == "-";
+            _hasInvariantNumberSigns = positiveSign == "+" && negativeSign == "-";
         }
 
         internal NumberFormatInfo(CultureData cultureData)
@@ -186,11 +178,6 @@ namespace System.Globalization
                 // don't need to verify their values (except for invalid parsing situations).
                 cultureData.GetNFIValues(this);
 
-                if (cultureData.IsInvariantCulture)
-                {
-                    // For invariant culture
-                    this.m_isInvariant = true;
-                }
                 UpdateHasInvariantNumberSigns();
             }
         }
@@ -216,9 +203,10 @@ namespace System.Globalization
                 {
                     // Lazy create the invariant info. This cannot be done in a .cctor because exceptions can
                     // be thrown out of a .cctor stack that will need this.
-                    NumberFormatInfo nfi = new NumberFormatInfo();
-                    nfi.m_isInvariant = true;
-                    s_invariantInfo = ReadOnly(nfi);
+                    s_invariantInfo = new NumberFormatInfo
+                    {
+                        isReadOnly = true
+                    };
                 }
                 return s_invariantInfo;
             }