… src/System.Private.CoreLib/shared/System/Number.Parsing.cs
… src/System.Private.CoreLib/shared/System/ParseNumbers.cs
… src/System.Private.CoreLib/shared/System/UInt16.cs
… src/System.Private.CoreLib/shared/System/UInt32.cs
… src/System.Private.CoreLib/shared/System/UInt64.cs
… src/System.Private.CoreLib/src/System/Enum.cs