As part of #19551 I tried to move byte parsing also to use UInt32 instead of Int32 for parsing, but that revealed some unexpected perf differences. Out of that investigation came this PR that speeds up common number parsing variants by 1.5x - 2x.

### x64
|         Method   |     Old  |    New   |
|-----------------:|---------:|---------:|
|     int.Parse(G) | 37.92 ns | 20.46 ns |
|  int.TryParse(G) | 36.69 ns | 20.66 ns |
| uint.TryParse(X) | 31.69 ns | 19.97 ns |
| byte.TryParse(X) | 28.02 ns | 14.82 ns |
|sbyte.TryParse(X) | 26.46 ns | 13.68 ns |
|    byte.Parse(X) | 26.30 ns | 12.14 ns |

### x86
|         Method   |     Old  |    New   |
|-----------------:|---------:|---------:|
|     int.Parse(G) | 40.43 ns | 20.58 ns |
|  int.TryParse(G) | 38.40 ns | 21.68 ns |
| uint.TryParse(X) | 29.87 ns | 18.98 ns |
| byte.TryParse(X) | 24.68 ns | 12.21 ns |
|sbyte.TryParse(X) | 24.63 ns | 13.30 ns |
|    byte.Parse(X) | 25.96 ns | 12.83 ns |